refactor: share webrtc ingest core, add H.264 negotiation
